now added

+++ JT/code.php
@@ -42,25 +42,28 @@ if (isset($_POST['addUser'])) {
 
 <?php
 include('config/dbcon.php');
-
+if (isset($_SESSION['auth_user'])) {
 if (isset($_POST['addfeedback'])) {
-    $name = $_POST['name'];
-    $date = $_POST['date'];
+    
+    $date = date("Y-m-d");
     $message = $_POST['message'];
-    $user_id = $_POST['auth'];
+    $user_id = $_SESSION['auth_user']['user_id'];
 
     $user_query = "INSERT INTO feedback (user_ID,f_date,f_description) VALUES ('$user_id','$date','$message')";
     $user_query_run = mysqli_query($con, $user_query);
 
     if ($user_query_run) {
-        $_SESSION['status'] = "User Added Successfully";
+        $_SESSION['status'] = "Sent your";
         header("Location: feedback.php");
     } else {
-        $_SESSION['status'] = "User Registration Failed";
+        $_SESSION['status'] = "Failed to sending";
         header("Location: feedback.php");
     }
 }
+}
+else{
 
+}
 
 ?>
 <?php
@@ -169,9 +172,12 @@ if (isset($_SESSION['auth_user'])) {
 
         if ($user_query_run) {
 
+            $_SESSION['status'] = "Uniform added to the cart";
             header("Location: uniform-single.php");
+
         } else {
 
+            $_SESSION['status'] = "Uniform added Failed";
             header("Location: uniform-single.php");
         }
     }
@@ -204,8 +210,8 @@ if (isset($_SESSION['auth_user'])) {
         $stmt->bind_param('i', $id);
         $stmt->execute();
 
-        $_SESSION['showAlert'] = 'block';
-        $_SESSION['message'] = 'Item removed from the cart!';
+        $_SESSION['status'] = "Item removed from the cart!";
+
         header('location:cart.php');
     }
 
@@ -222,7 +228,7 @@ if (isset($_SESSION['auth_user'])) {
         $user_query = "UPDATE orderdetails SET quantity= $qty,total_price=$tprice WHERE order_ID= $o_id";
         $user_query_run = mysqli_query($con, $user_query);
         if ($user_query_run) {
-
+            $_SESSION['status'] = "Item quatity added successfully!";
             header("Location: cart.php");
         }
     }
@@ -245,7 +251,7 @@ if (isset($_SESSION['auth_user'])) {
         $stmt = $con->prepare('INSERT INTO orders (user_ID,orders,order_date,due_date,delivery_address,district,postal_code)VALUES(?,?,?,?,?,?,?)');
         $stmt->bind_param('sssssss', $user_id, $products, $order_date, $due_date, $address, $district, $postal);
         $stmt->execute();
-        $stmt2 = $con->prepare('DELETE FROM orderdetails');
+        $stmt2 = $con->prepare("DELETE FROM orderdetails WHERE user_ID = '$user_id'");
         $stmt2->execute();
         $data .= '<div class="site-section">
                     <div class="container">
@@ -261,5 +267,341 @@ if (isset($_SESSION['auth_user'])) {
                 </div>';
         echo $data;
     }
-    ?>
+?>
+
+
+<?php
+include('config/dbcon.php');
+if (isset($_POST['AddCuniform'])) {
+
+
+    $uni_institition = $_POST['institution'];
+    $uni_color = $_POST['color'];
+    $uni_fabric_type = $_POST['fabric_type'];
+    $uni_extra_note = $_POST['extra_note'];
+    $img = $_POST['img'];
+
+    $qty = $_POST['c_qty'];
+    $category = $_POST['messagetype'];
+    $time = date("G:i:s A");
+    $date = date("M/d/Y");
+
+    $date_time = $time. ''.$date;
+
+    $user_id = $_SESSION['auth_user']['user_id'];
+
+    $selected_size = 'Can be change';
+
+    if($_POST['messagetype'] == 'Skirt'){
+
+        $chest = $_POST['ski_chest'];
+        $waist = $_POST['ski_waist'];
+        $full_length = $_POST['ski_full_length'];
+        $hip = $_POST['ski_hip'];
+
+
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_Type,extra_note,design_image,chest,waist,full_length,hip,date_time) VALUES ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$chest','$waist','$full_length','$hip','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+            
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'ski_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $full_length);
+
+            $total_price = $price * $qty;
+
+            $selected_size = 'Can be change';
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        } 
+
+    }
+    else if($_POST['messagetype'] == 'Pre school frock'){
+        
+        $ps_chest = $_POST['ps_chest'];
+        $ps_waist = $_POST['ps_waist'];
+        $ps_full_length = $_POST['ps_full_length'];
+        $ps_hip = $_POST['ps_hip'];
+        $ps_shoulder_to_waist = $_POST['ps_shoulder_to_waist'];
+        $ps_arm_hole = $_POST['ps_arm_hole'];
+        $ps_sleeve_circumference = $_POST['ps_sleeve_circumference'];
+        $ps_sleeve_length = $_POST['ps_sleeve_length'];
+        $ps_shoulder_length = $_POST['ps_shoulder_length'];
+
+
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_type,extra_note,design_image,chest,waist,full_length,hip,shoulder_length,sleeve_length,sleeve_circumference,arm_hole,shoulder_to_waist,date_time) VALUES ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$ps_chest','$ps_waist','$ps_full_length','$ps_hip','$ps_shoulder_length','$ps_sleeve_length','$ps_sleeve_circumference','$ps_arm_hole','$ps_shoulder_to_waist','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'ps_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $ps_full_length);
+
+            $total_price = $price * $qty;
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        }
+
+    }
+    else if($_POST['messagetype'] == 'Shirt'){
+
+        $shi_chest = $_POST['shi_chest'];
+        $shi_waist = $_POST['shi_waist'];
+        $shi_full_length = $_POST['shi_full_length'];
+        $shi_shoulder_length = $_POST['shi_shoulder_length'];
+        $shi_sleeve_length = $_POST['shi_sleeve_length'];
+        $shi_sleeve_circumference = $_POST['shi_sleeve_circumference'];
+        $shi_arm_hole = $_POST['shi_arm_hole'];
+        $shi_shoulder_to_waist = $_POST['shi_shoulder_to_waist'];
+        
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_type,extra_note,design_image,chest,waist,full_length,shoulder_length,sleeve_length,sleeve_circumference,arm_hole,shoulder_to_waist,date_time) VALUES 
+        ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$shi_chest','$shi_waist','$shi_full_length','$shi_shoulder_length','$shi_sleeve_length','$shi_sleeve_circumference','$shi_arm_hole','$shi_shoulder_to_waist','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'shi_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $shi_full_length);
+
+            $total_price = $price * $qty;
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        }
+
+    }
+    else if($_POST['messagetype'] == 'School frock'){
+        
+        $sch_chest = $_POST['sch_chest'];
+        $sch_waist = $_POST['sch_waist'];
+        $sch_full_length = $_POST['sch_full_length'];
+        $sch_hip = $_POST['sch_hip'];
+        $sch_shoulder_to_waist = $_POST['sch_shoulder_to_waist'];
+        $sch_arm_hole = $_POST['sch_arm_hole'];
+        $sch_sleeve_circumference = $_POST['sch_sleeve_circumference'];
+        $sch_sleeve_length = $_POST['sch_sleeve_length'];
+        $sch_shoulder_length = $_POST['sch_shoulder_length'];
+
+
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_type,extra_note,design_image,chest,waist,full_length,hip,shoulder_length,sleeve_length,sleeve_circumference,arm_hole,shoulder_to_waist,date_time) VALUES ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$sch_chest','$sch_waist','$sch_full_length','$sch_hip','$sch_shoulder_length','$sch_sleeve_length','$sch_sleeve_circumference','$sch_arm_hole','$sch_shoulder_to_waist','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'sch_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $sch_full_length);
+
+            $total_price = $price * $qty;
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        }
+
+    }
+    else if($_POST['messagetype'] == 'Short'){
+        
+        
+        $sho_waist = $_POST['sho_waist'];
+        $sho_full_length = $_POST['sho_full_length'];
+        $sho_hip = $_POST['sho_hip'];
+        $sho_front_crotch = $_POST['sho_front_crotch'];
+        $sho_bottom = $_POST['sho_bottom'];
+
+
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_type,extra_note,design_image,waist,full_length,hip,front_crotch,bottom,date_time) VALUES 
+        ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$sho_waist','$sho_full_length','$sho_hip','$sho_front_crotch','$sho_bottom','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'sho_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $sho_full_length);
+
+            $total_price = $price * $qty;
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        }
+
+    }
+    else if($_POST['messagetype'] == 'Trouser'){
+        
+        
+        $tro_waist = $_POST['tro_waist'];
+        $tro_full_length = $_POST['tro_full_length'];
+        $tro_hip = $_POST['tro_hip'];
+        $tro_front_crotch = $_POST['tro_front_crotch'];
+        $tro_bottom = $_POST['tro_bottom'];
+
+
+        $user_query = "INSERT INTO customizeduniform (institution,color,fabric_type,extra_note,design_image,waist,full_length,hip,front_crotch,bottom,date_time) VALUES 
+        ('$uni_institition','$uni_color','$uni_fabric_type','$uni_extra_note','$img','$tro_waist','$tro_full_length','$tro_hip','$tro_front_crotch','$tro_bottom','$date_time')";
+        $user_query_run = mysqli_query($con, $user_query);
+
+        if ($user_query_run) {
+
+            $query = "SELECT * FROM categoryprice WHERE c_name  = '$category'";
+            $result = $con->query($query);
+            $row = mysqli_fetch_assoc($result);
+            $avg_price = $row['avg_price'];
+
+            $query1 = "SELECT * FROM fabric WHERE fabric_Type  = '$uni_fabric_type'";
+            $result1 = $con->query($query1);
+            $row1 = mysqli_fetch_assoc($result1);
+            $extra_cost	 = $row1['extra_cost'];
+
+            $query2 = "SELECT * FROM customizeduniform WHERE date_time  = '$date_time'";
+            $result2 = $con->query($query2);
+            $row2 = mysqli_fetch_assoc($result2);
+            $c_uni_ID = $row2['cu_ID'];
+
+            $c_uniform_ID = 'tro_' .$c_uni_ID;
+
+            $price = $avg_price + ($extra_cost * $tro_full_length);
+
+            $total_price = $price * $qty;
+            
+            $user_query = "INSERT INTO orderdetails (uniform_ID,user_ID,selected_size,price,quantity,total_price) VALUES ('$c_uniform_ID','$user_id','$selected_size','$price','$qty','$total_price')";
+            $user_query_run = mysqli_query($con, $user_query);
+
+            $_SESSION['status'] = "Uniform added to the cart!";
+            header("Location: customizeduniform.php");
+        } 
+        else 
+        {
+            $_SESSION['status'] = "Uniform adding failed";
+            header("Location: customizeduniform.php");
+        }
+
+    }
+
+    
+
+}
+?>
+
+
+
